+++ argos_bridge/__init__.py
@@ -1,3 +1,3 @@
-"""ARGOS-bridge package."""
+"""University Report Governance System package."""
 
 __version__ = "0.2.0"